third_party/blink/web_tests/flag-specific/highdpi/animations/reverse-transition-cross-browser-expected.txt

This is a testharness.js-based test.
PASS Cross-browser check that reversing a transition mid-way adjusts the duration
FAIL Cross-browser check that reversing a transition with a non linear timing function mid-way adjusts the duration assert_approx_equals: expected 166.656 +/- 0.01 but got 166.667
Harness: the test ran to completion.
